Move Validator to package object casper
