+++ casper/src/main/scala/coop/rchain/casper/api/GraphGenerator.scala
@@ -1,7 +1,7 @@
 package coop.rchain.casper.api
 
 import coop.rchain.blockstorage.{BlockDagRepresentation, BlockStore}
-import coop.rchain.casper._, Estimator.BlockHash, MultiParentCasperRef.MultiParentCasperRef
+import coop.rchain.casper._, MultiParentCasperRef.MultiParentCasperRef
 import coop.rchain.casper.util.ProtoUtil
 import coop.rchain.graphz._
 import coop.rchain.shared.Log
