+++ casper/src/main/scala/coop/rchain/casper/package.scala
@@ -6,6 +6,7 @@ import coop.rchain.metrics.Metrics
 package object casper {
 
   type BlockHash = ByteString
+  type Validator = ByteString
 
   val CasperMetricsSource: Metrics.Source = Metrics.Source(Metrics.BaseSource, "casper")
 }
