+++ casper/src/main/scala/coop/rchain/casper/util/ProtoUtil.scala
@@ -4,9 +4,8 @@ import cats.{Applicative, Monad}
 import cats.implicits._
 import com.google.protobuf.{ByteString, Int32Value, StringValue}
 import coop.rchain.blockstorage.{BlockDagRepresentation, BlockStore}
-import coop.rchain.casper.PrettyPrinter
+import coop.rchain.casper.{BlockHash, PrettyPrinter, Validator}
 import coop.rchain.models.EquivocationRecord.SequenceNumber
-import coop.rchain.casper.Estimator.{BlockHash, Validator}
 import coop.rchain.casper.protocol.{DeployData, _}
 import coop.rchain.casper.util.rholang.InterpreterUtil
 import coop.rchain.casper.util.implicits._
