+++ casper/src/test/scala/coop/rchain/casper/HashSetCasperTest.scala
@@ -8,7 +8,6 @@ import cats.effect.{Concurrent, Sync}
 import cats.implicits._
 import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.BlockStore
-import coop.rchain.casper.Estimator.Validator
 import coop.rchain.casper.genesis.Genesis
 import coop.rchain.casper.genesis.contracts._
 import coop.rchain.casper.MultiParentCasper.ignoreDoppelgangerCheck
