+++ casper/src/test/scala/coop/rchain/casper/api/BlockQueryResponseAPITest.scala
@@ -8,7 +8,6 @@ import cats.implicits._
 import coop.rchain.crypto.{PrivateKey, PublicKey}
 import coop.rchain.blockstorage.{BlockDagStorage, BlockStore}
 import coop.rchain.casper._
-import coop.rchain.casper.Estimator.BlockHash
 import coop.rchain.casper.MultiParentCasperRef.MultiParentCasperRef
 import coop.rchain.casper.helper.{BlockDagStorageFixture, NoOpsCasperEffect}
 import coop.rchain.casper.protocol._
