+++ casper/src/test/scala/coop/rchain/casper/api/CreateBlockAPITest.scala
@@ -13,7 +13,6 @@ import coop.rchain.casper.helper.HashSetCasperTestNode
 import coop.rchain.casper.protocol._
 import coop.rchain.casper.util._
 import coop.rchain.casper.util.rholang._
-import coop.rchain.casper.Estimator.Validator
 import coop.rchain.casper.MultiParentCasper.ignoreDoppelgangerCheck
 import coop.rchain.casper.MultiParentCasperRef.MultiParentCasperRef
 import coop.rchain.casper.api.BlockAPI.ApiErr
