+++ casper/src/test/scala/coop/rchain/casper/helper/BlockGenerator.scala
@@ -5,7 +5,7 @@ import cats.effect._
 import cats.implicits._
 import com.google.protobuf.ByteString
 import coop.rchain.blockstorage._
-import coop.rchain.casper.Estimator.{BlockHash, Validator}
+import coop.rchain.casper.{BlockHash, Validator}
 import coop.rchain.casper.protocol._
 import coop.rchain.casper.util.ProtoUtil
 import coop.rchain.casper.util.rholang.{InterpreterUtil, ProcessedDeployUtil, RuntimeManager}
